requirements.txt
asana==0.8.2
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
Deprecated==1.2.5
dj-database-url==0.5.0
Django==2.2.1
django-filter==2.1.0
django-heroku==0.3.1
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
httpie==1.0.2
idna==2.8
ipaddress==1.0.22
Markdown==3.1.1
oauthlib==3.0.1
Pillow==6.0.0
psycopg2==2.8.2
PyGithub==1.43.7
Pygments==2.4.2
PyJWT==1.7.1
pytz==2019.1
requests==2.22.0
requests-oauthlib==0.8.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.2
wrapt==1.11.2
